XC-412: Move `AlloyResponseConverter` to separate file
